+++ src/TestNeuron.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 public class TestNeuron {
 
     public static void main(String[] args) {
-        float threshold = 0;
+        double threshold = 0;
 
 
         InputNeuron i1 = new InputNeuron(1);//A
@@ -57,8 +57,8 @@ public class TestNeuron {
         o2.assignWeight(-1, 1);//Assign a weight value of -1 to index 1
 
 
-        float result1 = o1.compute();
-        float result2 = o2.compute();
+        double result1 = o1.compute();
+        double result2 = o2.compute();
 
         NeuralNet net = new NeuralNet();
         net.addInNeuron(i1);
@@ -69,8 +69,8 @@ public class TestNeuron {
         net.addOutNeuron(o1);
         net.addOutNeuron(o2);
         Vector input = new Vector();
-        input.add(1f);
-        input.add(1f);
+        input.add(1.0);
+        input.add(1.0);
         System.out.println("Result 1 (Neural Net): "+net.calculate(input).get(0) + "," + "Result 2 (Neural Net): " + net.calculate(input).get(1));
         System.out.println("Result 1: " + result1 + " Result 2: " + result2 );
     }
