+++ src/mathematics/Vector.java
@@ -9,14 +9,14 @@ import java.util.ArrayList;
 /**
  * Vector implementation
  */
-public class Vector extends ArrayList<Float> {
+public class Vector extends ArrayList<Double> {
     /**
      * Computes the dot product between two vectors
      * @param v vector one
-     * @return a float value of the dot product
+     * @return a double value of the dot product
      */
-    public float dotProduct(Vector v){
-        float dotProduct = 0;
+    public double dotProduct(Vector v){
+        double dotProduct = 0;
         for(int i = 0; i < v.size(); ++i){
             dotProduct += v.get(i) * this.get(i);
         }
